requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Embeddings
sentence-transformers==2.2.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utils
tqdm==4.66.1
pyyaml==6.0.1

# Hugging Face
huggingface-hub>=0.20.0

# Streamlit
streamlit==1.29.0